[move]: move viewmodel file to directory
