+++ lib/src/views/onboarding/onboarding_view.dart
@@ -4,8 +4,8 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:green_field/src/cores/error_handler/result.dart';
 import 'package:green_field/src/datas/repositories/onboarding_repository.dart';
-import 'package:green_field/src/viewmodels/login_view_model.dart';
-import 'package:green_field/src/viewmodels/onboarding_view_model.dart';
+import 'package:green_field/src/viewmodels/login/login_view_model.dart';
+import 'package:green_field/src/viewmodels/onboarding/onboarding_view_model.dart';
 import 'package:kakao_flutter_sdk_user/kakao_flutter_sdk_user.dart';
 import '../../datas/services/firebase_auth_service.dart';
 import '../../utilities/components/greenfield_campus_picker.dart';
